refactor: move all subprocess-related code to new file
